add func format_cart, handle_cart

+++ tg_bot.py
@@ -11,13 +11,14 @@ from telegram.ext import (
 import redis
 
 from motlin_api import (
-    get_products, get_access_token, get_element_by_id, get_link_image)
+    get_products, get_access_token, get_element_by_id,
+    get_link_image, add_to_cart, get_cart)
 
 
 logger = logging.getLogger(__name__)
 
 
-def format_message(product_info):
+def format_description(product_info):
     name = product_info['name']
     description = product_info['description']
     price = product_info['meta']['display_price']['with_tax']['formatted']
@@ -34,13 +35,39 @@ def format_message(product_info):
     return textwrap.dedent(text_mess)
 
 
+def format_cart(cart):
+    filtred_cart = []
+    for fish in cart['data']:
+        filtred_cart.append({
+            'name': fish['name'],
+            "description": fish['description'],
+            'price_per_kg': fish['meta']['display_price']['without_tax']['unit']['formatted'],
+            'total': fish['meta']['display_price']['without_tax']['value']['formatted'],
+            'quantity': fish['quantity']
+
+        })
+    total_to_pay = cart['meta']['display_price']['without_tax']['formatted']
+    text_message = ''
+    for fish in filtred_cart:
+        text_message += (
+            f'''\
+
+            {fish["name"]}
+            {fish['description']}
+            {fish['price_per_kg']} price per kg
+            {fish['quantity']}kg in cart for {fish['total']}
+            ''')
+    text_message += total_to_pay
+    return textwrap.dedent(text_message)
+
+
 def start(redis_conn, update: Update, context: CallbackContext):
     keyboard = []
     access_token = get_access_token(redis_conn)
     for product in get_products(access_token):
         keyboard.append([InlineKeyboardButton(
             product['name'], callback_data=product['id'])])
-
+    keyboard.append([InlineKeyboardButton('Корзина', callback_data='Корзина')])
     reply_markup = InlineKeyboardMarkup(keyboard)
 
     context.bot.send_message(
@@ -49,29 +76,60 @@ def start(redis_conn, update: Update, context: CallbackContext):
     return 'HANDLE_MENU'
 
 
+def handle_cart(redis_conn, update: Update, context: CallbackContext):
+    access_token = get_access_token(redis_conn)
+    chat_id = update.effective_user.id
+    cart = get_cart(access_token, chat_id)
+    text_message = format_cart(cart)
+    keyboard = [
+        [InlineKeyboardButton('Назад', callback_data='Назад')]]
+    reply_markup = InlineKeyboardMarkup(keyboard)
+    context.bot.send_message(text=text_message, chat_id=chat_id, reply_markup=reply_markup)
+    return "HANDLE_DESCRIPTION"
+
+
 def handle_description(redis_conn, update: Update, context: CallbackContext):
     query = update.callback_query
     query.answer()
-    query.message.delete()
-    start(redis_conn, update, context)
-    return 'HANDLE_MENU'
+    if query.data == 'Назад':
+        query.message.delete()
+        start(redis_conn, update, context)
+        return 'HANDLE_MENU'
+    elif query.data == 'Корзина':
+        handle_cart(redis_conn, update, context)
+        return "HANDLE_DESCRIPTION"
+    quantity, item_id = query.data.split('|')
+    access_token = get_access_token(redis_conn)
+    add_to_cart(access_token, quantity, item_id, update.effective_user.id)
+    context.bot.send_message(
+        text='added', chat_id=update.effective_user.id)
+    return 'HANDLE_DESCRIPTION'
 
 
 def handle_menu(redis_conn, update: Update, context: CallbackContext):
     query = update.callback_query
     query.answer()
+    query.message.delete()
+    if query.data == 'Корзина':
+        handle_cart(redis_conn, update, context)
+        return "HANDLE_DESCRIPTION"
     access_token = get_access_token(redis_conn)
     product_info = get_element_by_id(access_token, query.data)
     image_id = product_info['relationships']['main_image']['data']['id']
-    text_mess = format_message(product_info)
+    text_mess = format_description(product_info)
     image_link = get_link_image(access_token, image_id)['data']['link']['href']
-    keyboard = [[InlineKeyboardButton('Назад', callback_data='Назад')]]
+    keyboard = [
+        [InlineKeyboardButton('Назад', callback_data='Назад')],
+        [
+            InlineKeyboardButton('1 кг', callback_data=f'1|{query.data}'),
+            InlineKeyboardButton('5 кг', callback_data=f'5|{query.data}'),
+            InlineKeyboardButton('10 кг', callback_data=f'10|{query.data}')
+            ],
+        [InlineKeyboardButton('Корзина', callback_data='Корзина')]]
     reply_markup = InlineKeyboardMarkup(keyboard)
-    chat_id = update.effective_user.id
-    message_id = context.bot.send_photo(
-        chat_id=chat_id, photo=image_link,
+    context.bot.send_photo(
+        chat_id=update.effective_user.id, photo=image_link,
         caption=text_mess, reply_markup=reply_markup)['message_id']
-    query.message.delete()
     return "HANDLE_DESCRIPTION"
 
 
@@ -85,6 +143,7 @@ def handle_users_reply(redis_conn, update: Update, context: CallbackContext):
     p_start = partial(start, redis_conn)
     p_handle_menu = partial(handle_menu, redis_conn)
     p_handle_description = partial(handle_description, redis_conn)
+    p_handle_cart = partial(handle_cart, redis_conn)
     if update.message:
         user_reply = update.message.text
     elif update.callback_query:
@@ -101,6 +160,7 @@ def handle_users_reply(redis_conn, update: Update, context: CallbackContext):
         'ECHO': echo,
         'HANDLE_MENU': p_handle_menu,
         'HANDLE_DESCRIPTION': p_handle_description,
+        'HANDLE_CART': p_handle_cart,
     }
     state_handler = states_functions[user_state]
     # Если вы вдруг не заметите, что python-telegram-bot перехватывает ошибки.
